requirements.txt
alembic==1.13.2
bcrypt==4.2.0
blinker==1.8.2
blueprint==3.4.2
boto3==1.35.1
botocore==1.35.1
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
flask-form==1.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-S3==0.3.3
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
